refactor(devtools): split CLI workbench modules
