+++ packages/devtools/src/cli-attached-session.ts
@@ -1,220 +1,52 @@
-import type { ChildProcess } from "node:child_process";
 import { spawn as defaultSpawn } from "node:child_process";
-import { types as nodeTypes } from "node:util";
 
-export const ATTACHED_CLI_SESSION_LIMITS = Object.freeze({
-  listTimeoutMs: 15_000,
-  describeTimeoutMs: 15_000,
-  runTimeoutMs: 60_000,
-  streamBytes: 10 * 1024 * 1024,
-  catalogSummaries: 2_000,
-  inputArgumentBytes: 98_304,
-  activityRecords: 500,
-  retainedActivityRecords: 50,
-  displayedNameCodePoints: 256,
-});
+import {
+  ATTACHED_CLI_SESSION_LIMITS,
+  attachedCliError,
+  type AttachedCliActivityRecord,
+  type AttachedCliCapabilityDescription,
+  type AttachedCliCapabilitySummary,
+  type AttachedCliChildResult,
+  type AttachedCliConnectionSummary,
+  type AttachedCliSessionClock,
+  type AttachedCliSessionController,
+  AttachedCliSessionError,
+  type AttachedCliSessionErrorCode,
+  type AttachedCliSpawn,
+  type CreateAttachedCliSessionControllerOptions,
+  type ParsedCliTarget,
+} from "./cli-attached-contract.js";
+import {
+  collectAttachedCliChild,
+  runAttachedCliWithDeadline,
+} from "./cli-attached-process.js";
+import {
+  encodeAttachedCliRunInput,
+  parseAttachedCliCatalog,
+  parseAttachedCliDescription,
+  parseAttachedCliJson,
+} from "./cli-attached-protocol.js";
+import {
+  composeAttachedCliEnvironment,
+  parseAttachedCliTarget,
+} from "./cli-attached-target.js";
 
-export type AttachedCliSessionErrorCode =
-  | "INVALID_TARGET"
-  | "SPAWN_FAILED"
-  | "CONNECTION_FAILED"
-  | "PROTOCOL_ERROR"
-  | "TIMEOUT"
-  | "LIMIT_EXCEEDED"
-  | "TARGET_BUSY"
-  | "NOT_CONNECTED"
-  | "ENVIRONMENT_VALUE_MISSING";
-
-const errorMessages = {
-  INVALID_TARGET: "The CLI target descriptor is invalid.",
-  SPAWN_FAILED: "The CLI process could not be started.",
-  CONNECTION_FAILED: "The CLI connection failed.",
-  PROTOCOL_ERROR: "The CLI returned an invalid document.",
-  TIMEOUT: "The CLI operation timed out.",
-  LIMIT_EXCEEDED: "The CLI operation exceeded a configured limit.",
-  TARGET_BUSY: "Another target or CLI verb is already active.",
-  NOT_CONNECTED: "No CLI target is connected.",
-  ENVIRONMENT_VALUE_MISSING: "A required environment value is missing.",
-} as const satisfies Record<AttachedCliSessionErrorCode, string>;
-
-const unixDefaultEnvNames = [
-  "HOME",
-  "LOGNAME",
-  "PATH",
-  "SHELL",
-  "TERM",
-  "USER",
-] as const;
-
-const windowsDefaultEnvNames = [
-  "APPDATA",
-  "HOMEDRIVE",
-  "HOMEPATH",
-  "LOCALAPPDATA",
-  "PATH",
-  "PROCESSOR_ARCHITECTURE",
-  "SYSTEMDRIVE",
-  "SYSTEMROOT",
-  "TEMP",
-  "USERNAME",
-  "USERPROFILE",
-  "PROGRAMFILES",
-] as const;
-
-const environmentNamePattern = /^[A-Za-z_][A-Za-z0-9_]*$/;
-const annotationKeys = [
-  "readOnly",
-  "destructive",
-  "idempotent",
-  "openWorld",
-] as const;
-const defaultKillGraceMs = 3_000;
-const deadlineReason = Object.freeze({ type: "attached-cli-deadline" });
-const disconnectReason = Object.freeze({ type: "attached-cli-disconnect" });
-
-export class AttachedCliSessionError extends Error {
-  declare readonly code: AttachedCliSessionErrorCode;
-  declare readonly message: string;
-
-  constructor(code: AttachedCliSessionErrorCode, options?: ErrorOptions) {
-    super();
-    Object.defineProperties(this, {
-      code: {
-        configurable: false,
-        enumerable: true,
-        value: code,
-        writable: false,
-      },
-      message: {
-        configurable: false,
-        enumerable: true,
-        value: errorMessages[code],
-        writable: false,
-      },
-      ...(options?.cause === undefined
-        ? {}
-        : {
-            cause: {
-              configurable: false,
-              enumerable: false,
-              value: options.cause,
-              writable: false,
-            },
-          }),
-    });
-  }
-}
-
-export interface AttachedCliAnnotations {
-  readonly readOnly?: boolean;
-  readonly destructive?: boolean;
-  readonly idempotent?: boolean;
-  readonly openWorld?: boolean;
-}
-
-export interface AttachedCliCapabilitySummary {
-  readonly id: string;
-  readonly description: string;
-  readonly title?: string;
-  readonly annotations?: AttachedCliAnnotations;
-}
-
-export interface AttachedCliCapabilityDescription
-  extends AttachedCliCapabilitySummary {
-  readonly inputSchema: Readonly<Record<string, unknown>>;
-  readonly outputSchema: Readonly<Record<string, unknown>>;
-  readonly timeoutMs?: number;
-}
-
-export interface AttachedCliActivityRecord {
-  readonly sequence: number;
-  readonly operation: "list" | "describe" | "run" | "disconnect";
-  readonly startedAt: string;
-  readonly durationMs: number;
-  readonly outcome: "success" | "error";
-  readonly errorCode?: AttachedCliSessionErrorCode;
-  readonly capabilityId?: string;
-  readonly exitCode?: number | null;
-}
-
-export interface AttachedCliConnectionSummary {
-  readonly command: string;
-  readonly capabilityCount: number;
-  readonly validation: { readonly status: "ok" };
-}
-
-export type AttachedCliSessionState =
-  | {
-      readonly state: "idle";
-      readonly validation?: {
-        readonly status: "error";
-        readonly error: {
-          readonly code: AttachedCliSessionErrorCode;
-          readonly message: string;
-        };
-      };
-      readonly activity?: readonly AttachedCliActivityRecord[];
-    }
-  | { readonly state: "busy" }
-  | { readonly state: "connecting" }
-  | {
-      readonly state: "connected";
-      readonly connection: AttachedCliConnectionSummary;
-    }
-  | { readonly state: "closing" };
-
-export interface AttachedCliSessionClock {
-  now(): number;
-  schedule(callback: () => void, delayMs: number): unknown;
-  cancel(handle: unknown): void;
-}
-
-export interface AttachedCliSpawnOptions {
-  readonly cwd?: string;
-  readonly env: NodeJS.ProcessEnv;
-  readonly shell: false;
-  readonly stdio: readonly ["ignore", "pipe", "pipe"];
-}
-
-export type AttachedCliSpawn = (
-  command: string,
-  args: readonly string[],
-  options: AttachedCliSpawnOptions,
-) => ChildProcess;
-
-export interface CreateAttachedCliSessionControllerOptions {
-  readonly spawn?: AttachedCliSpawn;
-  readonly clock?: AttachedCliSessionClock;
-  readonly killGraceMs?: number;
-  readonly platform?: NodeJS.Platform;
-  readonly readHostEnv?: (name: string) => string | undefined;
-}
-
-export interface AttachedCliSessionController {
-  connect(
-    owner: string,
-    target: unknown,
-  ): Promise<AttachedCliConnectionSummary>;
-  refresh(owner: string): Promise<AttachedCliConnectionSummary>;
-  describe(
-    owner: string,
-    id: string,
-  ): Promise<AttachedCliCapabilityDescription>;
-  run(owner: string, id: string, input: unknown): Promise<unknown>;
-  state(owner: string): AttachedCliSessionState;
-  catalog(owner: string): readonly AttachedCliCapabilitySummary[];
-  description(owner: string): AttachedCliCapabilityDescription | undefined;
-  activity(owner: string): readonly AttachedCliActivityRecord[];
-  disconnect(owner: string): Promise<void>;
-  close(): Promise<void>;
-}
-
-interface ParsedCliTarget {
-  readonly command: string;
-  readonly args: readonly string[];
-  readonly cwd?: string;
-  readonly overlay: Readonly<Record<string, string>>;
-}
+export {
+  ATTACHED_CLI_SESSION_LIMITS,
+  type AttachedCliActivityRecord,
+  type AttachedCliAnnotations,
+  type AttachedCliCapabilityDescription,
+  type AttachedCliCapabilitySummary,
+  type AttachedCliConnectionSummary,
+  type AttachedCliSessionClock,
+  type AttachedCliSessionController,
+  AttachedCliSessionError,
+  type AttachedCliSessionErrorCode,
+  type AttachedCliSessionState,
+  type AttachedCliSpawn,
+  type AttachedCliSpawnOptions,
+  type CreateAttachedCliSessionControllerOptions,
+} from "./cli-attached-contract.js";
 
 interface ActivityStore {
   append(
@@ -248,27 +80,14 @@ interface RetainedActivity {
   readonly records: readonly AttachedCliActivityRecord[];
 }
 
-interface ChildResult {
-  readonly exitCode: number | null;
-  readonly stdout: Buffer;
-  readonly stderr: Buffer;
-}
+const defaultKillGraceMs = 3_000;
+const disconnectReason = Object.freeze({ type: "attached-cli-disconnect" });
 
 const defaultClock: AttachedCliSessionClock = {
   now: () => Date.now(),
   schedule: (callback, delayMs) => setTimeout(callback, delayMs),
   cancel: (handle) => clearTimeout(handle as ReturnType<typeof setTimeout>),
 };
-
-function attachedError(
-  code: AttachedCliSessionErrorCode,
-  cause?: unknown,
-): AttachedCliSessionError {
-  return new AttachedCliSessionError(
-    code,
-    cause === undefined ? undefined : { cause },
-  );
-}
 
 function createActivityStore(): ActivityStore {
   const records: AttachedCliActivityRecord[] = [];
@@ -290,16 +109,6 @@ function createActivityStore(): ActivityStore {
   };
 }
 
-function ownDataProperty(value: object, key: string): unknown {
-  const descriptor = Object.getOwnPropertyDescriptor(value, key);
-  if (descriptor === undefined || !("value" in descriptor)) return undefined;
-  return descriptor.value;
-}
-
-function isPlainRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
 function duration(clock: AttachedCliSessionClock, started: number): number {
   return Math.max(0, clock.now() - started);
 }
@@ -318,205 +127,6 @@ function boundedCapabilityId(id: string): string {
     .join("");
 }
 
-function defaultEnvironment(
-  platform: NodeJS.Platform,
-  readHostEnv: (name: string) => string | undefined,
-): Record<string, string> {
-  const names =
-    platform === "win32" ? windowsDefaultEnvNames : unixDefaultEnvNames;
-  const env = Object.create(null) as Record<string, string>;
-  for (const name of names) {
-    const value = readHostEnv(name);
-    if (typeof value !== "string" || value.startsWith("()")) continue;
-    Object.defineProperty(env, name, {
-      configurable: true,
-      enumerable: true,
-      value,
-      writable: true,
-    });
-  }
-  return env;
-}
-
-function composeEnvironment(
-  overlay: Readonly<Record<string, string>>,
-  platform: NodeJS.Platform,
-  readHostEnv: (name: string) => string | undefined,
-): Record<string, string> {
-  const env = defaultEnvironment(platform, readHostEnv);
-  for (const name of Object.keys(overlay)) {
-    Object.defineProperty(env, name, {
-      configurable: true,
-      enumerable: true,
-      value: overlay[name],
-      writable: true,
-    });
-  }
-  return env;
-}
-
-function parseTarget(value: unknown): ParsedCliTarget {
-  if (
-    typeof value !== "object" ||
-    value === null ||
-    nodeTypes.isProxy(value) ||
-    Array.isArray(value)
-  ) {
-    throw attachedError("INVALID_TARGET");
-  }
-  const command = ownDataProperty(value, "command");
-  if (typeof command !== "string" || command.trim() === "") {
-    throw attachedError("INVALID_TARGET");
-  }
-  const rawArgs = ownDataProperty(value, "args");
-  let args: readonly string[] = [];
-  if (rawArgs !== undefined) {
-    if (
-      !Array.isArray(rawArgs) ||
-      rawArgs.some((entry) => typeof entry !== "string")
-    ) {
-      throw attachedError("INVALID_TARGET");
-    }
-    args = rawArgs;
-  }
-  const cwd = ownDataProperty(value, "cwd");
-  if (cwd !== undefined && (typeof cwd !== "string" || cwd.trim() === "")) {
-    throw attachedError("INVALID_TARGET");
-  }
-  const rawEnv = ownDataProperty(value, "env");
-  const overlay = Object.create(null) as Record<string, string>;
-  if (rawEnv !== undefined) {
-    if (!isPlainRecord(rawEnv) || nodeTypes.isProxy(rawEnv)) {
-      throw attachedError("INVALID_TARGET");
-    }
-    for (const name of Object.keys(rawEnv)) {
-      if (!environmentNamePattern.test(name)) {
-        throw attachedError("INVALID_TARGET");
-      }
-      const entry = ownDataProperty(rawEnv, name);
-      if (typeof entry !== "string") throw attachedError("INVALID_TARGET");
-      if (entry === "") throw attachedError("ENVIRONMENT_VALUE_MISSING");
-      Object.defineProperty(overlay, name, {
-        configurable: true,
-        enumerable: true,
-        value: entry,
-        writable: true,
-      });
-    }
-  }
-  return {
-    command,
-    args,
-    ...(typeof cwd === "string" ? { cwd } : {}),
-    overlay,
-  };
-}
-
-function parseOneJson(buffer: Buffer): unknown {
-  let text: string;
-  try {
-    text = new TextDecoder("utf-8", { fatal: true }).decode(buffer);
-  } catch {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-  try {
-    return JSON.parse(text) as unknown;
-  } catch {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-}
-
-function parseAnnotations(value: unknown): AttachedCliAnnotations | undefined {
-  if (value === undefined) return undefined;
-  if (!isPlainRecord(value)) throw attachedError("PROTOCOL_ERROR");
-  const annotations: Record<string, boolean> = {};
-  for (const key of annotationKeys) {
-    const entry = ownDataProperty(value, key);
-    if (entry === undefined) continue;
-    if (typeof entry !== "boolean") throw attachedError("PROTOCOL_ERROR");
-    annotations[key] = entry;
-  }
-  return Object.keys(annotations).length === 0
-    ? {}
-    : (annotations as AttachedCliAnnotations);
-}
-
-function parseSummary(value: unknown): AttachedCliCapabilitySummary {
-  if (!isPlainRecord(value)) throw attachedError("PROTOCOL_ERROR");
-  const id = ownDataProperty(value, "id");
-  const description = ownDataProperty(value, "description");
-  if (typeof id !== "string" || id === "") {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-  if (typeof description !== "string") throw attachedError("PROTOCOL_ERROR");
-  const title = ownDataProperty(value, "title");
-  if (title !== undefined && typeof title !== "string") {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-  const annotations = parseAnnotations(ownDataProperty(value, "annotations"));
-  return {
-    id,
-    description,
-    ...(title === undefined ? {} : { title }),
-    ...(annotations === undefined ? {} : { annotations }),
-  };
-}
-
-function parseCatalog(buffer: Buffer): readonly AttachedCliCapabilitySummary[] {
-  const document = parseOneJson(buffer);
-  if (!Array.isArray(document)) throw attachedError("PROTOCOL_ERROR");
-  if (document.length > ATTACHED_CLI_SESSION_LIMITS.catalogSummaries) {
-    throw attachedError("LIMIT_EXCEEDED");
-  }
-  return Object.freeze(document.map((entry) => parseSummary(entry)));
-}
-
-function parseDescription(buffer: Buffer): AttachedCliCapabilityDescription {
-  const document = parseOneJson(buffer);
-  if (!isPlainRecord(document)) throw attachedError("PROTOCOL_ERROR");
-  const summary = parseSummary(document);
-  const inputSchema = ownDataProperty(document, "inputSchema");
-  const outputSchema = ownDataProperty(document, "outputSchema");
-  if (
-    !isPlainRecord(inputSchema) ||
-    !isPlainRecord(outputSchema) ||
-    Array.isArray(inputSchema) ||
-    Array.isArray(outputSchema)
-  ) {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-  const timeoutMs = ownDataProperty(document, "timeoutMs");
-  if (
-    timeoutMs !== undefined &&
-    (typeof timeoutMs !== "number" || !Number.isFinite(timeoutMs))
-  ) {
-    throw attachedError("PROTOCOL_ERROR");
-  }
-  return {
-    ...summary,
-    inputSchema,
-    outputSchema,
-    ...(timeoutMs === undefined ? {} : { timeoutMs }),
-  };
-}
-
-function encodeRunInput(input: unknown): string {
-  let encoded: string;
-  try {
-    encoded = JSON.stringify(input);
-  } catch {
-    throw attachedError("INVALID_TARGET");
-  }
-  if (typeof encoded !== "string") throw attachedError("INVALID_TARGET");
-  if (
-    Buffer.byteLength(encoded, "utf8") >
-    ATTACHED_CLI_SESSION_LIMITS.inputArgumentBytes
-  ) {
-    throw attachedError("LIMIT_EXCEEDED");
-  }
-  return encoded;
-}
-
 function connectionSummary(
   target: ParsedCliTarget,
   catalog: readonly AttachedCliCapabilitySummary[],
@@ -525,200 +135,6 @@ function connectionSummary(
     command: target.command,
     capabilityCount: catalog.length,
     validation: Object.freeze({ status: "ok" as const }),
-  });
-}
-
-function killChild(
-  child: ChildProcess,
-  clock: AttachedCliSessionClock,
-  graceMs: number,
-): void {
-  if (child.exitCode !== null || child.signalCode !== null) return;
-  child.kill("SIGTERM");
-  const handle = clock.schedule(() => {
-    if (child.exitCode === null && child.signalCode === null) {
-      child.kill("SIGKILL");
-    }
-  }, graceMs);
-  child.once("exit", () => {
-    clock.cancel(handle);
-  });
-}
-
-function collectChild(
-  spawn: AttachedCliSpawn,
-  target: ParsedCliTarget,
-  verbArgs: readonly string[],
-  env: Record<string, string>,
-  clock: AttachedCliSessionClock,
-  killGraceMs: number,
-  signal: AbortSignal,
-): Promise<ChildResult> {
-  return new Promise((resolve, reject) => {
-    let settled = false;
-    let overflow = false;
-    let child: ChildProcess;
-    try {
-      child = spawn(target.command, [...target.args, ...verbArgs], {
-        ...(target.cwd === undefined ? {} : { cwd: target.cwd }),
-        env,
-        shell: false,
-        stdio: ["ignore", "pipe", "pipe"],
-      });
-    } catch (cause) {
-      reject(attachedError("SPAWN_FAILED", cause));
-      return;
-    }
-
-    const stdoutChunks: Buffer[] = [];
-    const stderrChunks: Buffer[] = [];
-    let stdoutBytes = 0;
-    let stderrBytes = 0;
-
-    const finish = (
-      error?: AttachedCliSessionError,
-      result?: ChildResult,
-    ): void => {
-      if (settled) return;
-      settled = true;
-      signal.removeEventListener("abort", onAbort);
-      if (error !== undefined) reject(error);
-      else if (result !== undefined) resolve(result);
-    };
-
-    const onAbort = (): void => {
-      killChild(child, clock, killGraceMs);
-    };
-
-    const onChunk = (
-      stream: "stdout" | "stderr",
-      value: Buffer | string,
-    ): void => {
-      const chunk = Buffer.isBuffer(value) ? value : Buffer.from(value);
-      if (stream === "stdout") {
-        if (
-          stdoutBytes + chunk.length >
-          ATTACHED_CLI_SESSION_LIMITS.streamBytes
-        ) {
-          overflow = true;
-          killChild(child, clock, killGraceMs);
-          return;
-        }
-        stdoutChunks.push(chunk);
-        stdoutBytes += chunk.length;
-        return;
-      }
-      if (
-        stderrBytes + chunk.length >
-        ATTACHED_CLI_SESSION_LIMITS.streamBytes
-      ) {
-        overflow = true;
-        killChild(child, clock, killGraceMs);
-        return;
-      }
-      stderrChunks.push(chunk);
-      stderrBytes += chunk.length;
-    };
-
-    child.stdout?.on("data", (chunk: Buffer | string) => {
-      onChunk("stdout", chunk);
-    });
-    child.stderr?.on("data", (chunk: Buffer | string) => {
-      onChunk("stderr", chunk);
-    });
-    child.once("error", (cause) => {
-      finish(attachedError("SPAWN_FAILED", cause));
-    });
-    child.once("exit", (code) => {
-      if (overflow) {
-        finish(attachedError("LIMIT_EXCEEDED"));
-        return;
-      }
-      if (signal.aborted) {
-        finish(
-          attachedError(
-            signal.reason === deadlineReason ? "TIMEOUT" : "NOT_CONNECTED",
-          ),
-        );
-        return;
-      }
-      finish(undefined, {
-        exitCode: code,
-        stdout: Buffer.concat(stdoutChunks),
-        stderr: Buffer.concat(stderrChunks),
-      });
-    });
-    signal.addEventListener("abort", onAbort, { once: true });
-    if (signal.aborted) onAbort();
-  });
-}
-
-function runWithDeadline<Value>(
-  clock: AttachedCliSessionClock,
-  timeoutMs: number,
-  controller: AbortController,
-  operation: (signal: AbortSignal) => Promise<Value>,
-): Promise<Value> {
-  return new Promise<Value>((resolve, reject) => {
-    let settled = false;
-    let handle: unknown;
-    const cleanup = (): void => {
-      clock.cancel(handle);
-      controller.signal.removeEventListener("abort", onAbort);
-    };
-    const fail = (error: AttachedCliSessionError): void => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      reject(error);
-    };
-    const onAbort = (): void => {
-      fail(
-        attachedError(
-          controller.signal.reason === deadlineReason
-            ? "TIMEOUT"
-            : "NOT_CONNECTED",
-        ),
-      );
-    };
-    controller.signal.addEventListener("abort", onAbort, { once: true });
-    handle = clock.schedule(() => {
-      if (settled) return;
-      settled = true;
-      cleanup();
-      controller.abort(deadlineReason);
-      reject(attachedError("TIMEOUT"));
-    }, timeoutMs);
-
-    let pending: Promise<Value>;
-    try {
-      pending = operation(controller.signal);
-    } catch (cause) {
-      fail(
-        cause instanceof AttachedCliSessionError
-          ? cause
-          : attachedError("SPAWN_FAILED", cause),
-      );
-      return;
-    }
-    void pending.then(
-      (value) => {
-        if (settled) return;
-        settled = true;
-        cleanup();
-        resolve(value);
-      },
-      (cause: unknown) => {
-        if (settled) return;
-        settled = true;
-        cleanup();
-        reject(
-          cause instanceof AttachedCliSessionError
-            ? cause
-            : attachedError("CONNECTION_FAILED", cause),
-        );
-      },
-    );
   });
 }
 
@@ -798,11 +214,11 @@ export function createAttachedCliSessionController(
   };
 
   const requireOwner = (owner: string, connected: boolean): ActiveSlot => {
-    if (active === undefined) throw attachedError("NOT_CONNECTED");
-    if (active.owner !== owner) throw attachedError("TARGET_BUSY");
-    if (active.verbActive) throw attachedError("TARGET_BUSY");
+    if (active === undefined) throw attachedCliError("NOT_CONNECTED");
+    if (active.owner !== owner) throw attachedCliError("TARGET_BUSY");
+    if (active.verbActive) throw attachedCliError("TARGET_BUSY");
     if (connected && active.state !== "connected") {
-      throw attachedError("NOT_CONNECTED");
+      throw attachedCliError("NOT_CONNECTED");
     }
     return active;
   };
@@ -811,22 +227,30 @@ export function createAttachedCliSessionController(
     slot: ActiveSlot,
     verbArgs: readonly string[],
     timeoutMs: number,
-  ): Promise<ChildResult> => {
-    if (slot.verbActive) throw attachedError("TARGET_BUSY");
+  ): Promise<AttachedCliChildResult> => {
+    if (slot.verbActive) throw attachedCliError("TARGET_BUSY");
     slot.verbActive = true;
     const controller = new AbortController();
     slot.verbAbort = controller;
-    const env = composeEnvironment(slot.target.overlay, platform, readHostEnv);
-    const pending = runWithDeadline(clock, timeoutMs, controller, (signal) =>
-      collectChild(
-        spawn,
-        slot.target,
-        verbArgs,
-        env,
-        clock,
-        killGraceMs,
-        signal,
-      ),
+    const env = composeAttachedCliEnvironment(
+      slot.target.overlay,
+      platform,
+      readHostEnv,
+    );
+    const pending = runAttachedCliWithDeadline(
+      clock,
+      timeoutMs,
+      controller,
+      (signal) =>
+        collectAttachedCliChild(
+          spawn,
+          slot.target,
+          verbArgs,
+          env,
+          clock,
+          killGraceMs,
+          signal,
+        ),
     );
     slot.verbPromise = pending;
     try {
@@ -849,14 +273,14 @@ export function createAttachedCliSessionController(
         ATTACHED_CLI_SESSION_LIMITS.listTimeoutMs,
       );
       if (child.exitCode !== 0) {
-        const failure = attachedError("CONNECTION_FAILED");
+        const failure = attachedCliError("CONNECTION_FAILED");
         appendActivity(slot, "list", started, "error", {
           errorCode: failure.code,
           exitCode: child.exitCode,
         });
         throw failure;
       }
-      const catalog = parseCatalog(child.stdout);
+      const catalog = parseAttachedCliCatalog(child.stdout);
       appendActivity(slot, "list", started, "success", {
         exitCode: child.exitCode,
       });
@@ -865,7 +289,7 @@ export function createAttachedCliSessionController(
       const failure =
         cause instanceof AttachedCliSessionError
           ? cause
-          : attachedError("CONNECTION_FAILED", cause);
+          : attachedCliError("CONNECTION_FAILED", cause);
       if (
         !slot.activity
           .entries()
@@ -887,8 +311,8 @@ export function createAttachedCliSessionController(
     owner: string,
     target: unknown,
   ): Promise<AttachedCliConnectionSummary> => {
-    if (active !== undefined) throw attachedError("TARGET_BUSY");
-    const parsed = parseTarget(target);
+    if (active !== undefined) throw attachedCliError("TARGET_BUSY");
+    const parsed = parseAttachedCliTarget(target);
     lastValidationFailure = undefined;
     retainedActivity = undefined;
     const slot: ActiveSlot = {
@@ -906,7 +330,7 @@ export function createAttachedCliSessionController(
     active = slot;
     try {
       const catalog = await runList(slot);
-      if (active !== slot) throw attachedError("NOT_CONNECTED");
+      if (active !== slot) throw attachedCliError("NOT_CONNECTED");
       slot.catalog = catalog;
       slot.connectionSummary = connectionSummary(parsed, catalog);
       slot.state = "connected";
@@ -915,7 +339,7 @@ export function createAttachedCliSessionController(
       const failure =
         cause instanceof AttachedCliSessionError
           ? cause
-          : attachedError("CONNECTION_FAILED", cause);
+          : attachedCliError("CONNECTION_FAILED", cause);
       failConnection(slot, failure);
       throw failure;
     }
@@ -927,7 +351,7 @@ export function createAttachedCliSessionController(
     const slot = requireOwner(owner, true);
     try {
       const catalog = await runList(slot);
-      if (active !== slot) throw attachedError("NOT_CONNECTED");
+      if (active !== slot) throw attachedCliError("NOT_CONNECTED");
       slot.catalog = catalog;
       slot.described = undefined;
       slot.connectionSummary = connectionSummary(slot.target, catalog);
@@ -936,7 +360,7 @@ export function createAttachedCliSessionController(
       const failure =
         cause instanceof AttachedCliSessionError
           ? cause
-          : attachedError("CONNECTION_FAILED", cause);
+          : attachedCliError("CONNECTION_FAILED", cause);
       failConnection(slot, failure);
       throw failure;
     }
@@ -948,10 +372,10 @@ export function createAttachedCliSessionController(
   ): Promise<AttachedCliCapabilityDescription> => {
     const slot = requireOwner(owner, true);
     if (typeof id !== "string" || id === "") {
-      throw attachedError("INVALID_TARGET");
+      throw attachedCliError("INVALID_TARGET");
     }
     if (!slot.catalog.some((entry) => entry.id === id)) {
-      throw attachedError("PROTOCOL_ERROR");
+      throw attachedCliError("PROTOCOL_ERROR");
     }
     const started = clock.now();
     try {
@@ -961,7 +385,7 @@ export function createAttachedCliSessionController(
         ATTACHED_CLI_SESSION_LIMITS.describeTimeoutMs,
       );
       if (child.exitCode !== 0) {
-        const failure = attachedError("CONNECTION_FAILED");
+        const failure = attachedCliError("CONNECTION_FAILED");
         appendActivity(slot, "describe", started, "error", {
           errorCode: failure.code,
           capabilityId: id,
@@ -970,7 +394,7 @@ export function createAttachedCliSessionController(
         slot.described = undefined;
         throw failure;
       }
-      const described = parseDescription(child.stdout);
+      const described = parseAttachedCliDescription(child.stdout);
       appendActivity(slot, "describe", started, "success", {
         capabilityId: id,
         exitCode: child.exitCode,
@@ -981,7 +405,7 @@ export function createAttachedCliSessionController(
       const failure =
         cause instanceof AttachedCliSessionError
           ? cause
-          : attachedError("CONNECTION_FAILED", cause);
+          : attachedCliError("CONNECTION_FAILED", cause);
       if (
         !slot.activity
           .entries()
@@ -1008,9 +432,9 @@ export function createAttachedCliSessionController(
   ): Promise<unknown> => {
     const slot = requireOwner(owner, true);
     if (slot.described === undefined || slot.described.id !== id) {
-      throw attachedError("NOT_CONNECTED");
+      throw attachedCliError("NOT_CONNECTED");
     }
-    const encoded = encodeRunInput(input);
+    const encoded = encodeAttachedCliRunInput(input);
     const started = clock.now();
     try {
       const child = await runVerb(
@@ -1019,7 +443,7 @@ export function createAttachedCliSessionController(
         ATTACHED_CLI_SESSION_LIMITS.runTimeoutMs,
       );
       if (child.exitCode !== 0) {
-        const failure = attachedError("CONNECTION_FAILED");
+        const failure = attachedCliError("CONNECTION_FAILED");
         appendActivity(slot, "run", started, "error", {
           errorCode: failure.code,
           capabilityId: id,
@@ -1027,7 +451,7 @@ export function createAttachedCliSessionController(
         });
         throw failure;
       }
-      const result = parseOneJson(child.stdout);
+      const result = parseAttachedCliJson(child.stdout);
       appendActivity(slot, "run", started, "success", {
         capabilityId: id,
         exitCode: child.exitCode,
@@ -1037,7 +461,7 @@ export function createAttachedCliSessionController(
       const failure =
         cause instanceof AttachedCliSessionError
           ? cause
-          : attachedError("CONNECTION_FAILED", cause);
+          : attachedCliError("CONNECTION_FAILED", cause);
       if (
         !slot.activity
           .entries()
@@ -1059,7 +483,7 @@ export function createAttachedCliSessionController(
   const disconnect = async (owner: string): Promise<void> => {
     const slot = active;
     if (slot === undefined) return;
-    if (slot.owner !== owner) throw attachedError("TARGET_BUSY");
+    if (slot.owner !== owner) throw attachedCliError("TARGET_BUSY");
     slot.state = "closing";
     slot.verbAbort?.abort(disconnectReason);
     const started = clock.now();
@@ -1112,8 +536,9 @@ export function createAttachedCliSessionController(
       };
     },
     catalog(owner) {
-      if (active === undefined || active.owner !== owner)
+      if (active === undefined || active.owner !== owner) {
         return Object.freeze([]);
+      }
       return active.catalog;
     },
     description(owner) {
