+++ packages/devtools/src/ui/cli-app.ts
@@ -1,538 +1,60 @@
-import { createCopyButton, formatDuration } from "./clipboard.js";
-import { clear, el, pretty } from "./dom.js";
-import { exampleFromSchema } from "./example-from-schema.js";
+import { CliApiError, createRouteCliApi } from "./cli-api.js";
+import {
+  type CliCurrentResult,
+  createCliCommandsPanel,
+} from "./cli-command-view.js";
+import {
+  cliActivityTable,
+  cliErrorMessage,
+  cliStatusPill,
+  createCliBrandMark,
+} from "./cli-components.js";
+import {
+  type CliActivityRecord,
+  type CliApi,
+  type CliCapabilityDescription,
+  type CliCapabilitySummary,
+  type CliConnectionState,
+  refreshFailureIsDisconnect,
+  retainedActivityOf,
+} from "./cli-contract.js";
+import {
+  createCliIdleView,
+  createCliUnavailableView,
+  type EditableCliTargetDraft,
+} from "./cli-connection-view.js";
+import { el } from "./dom.js";
 import { createCompactThemeToggle, createThemeToggle } from "./theme.js";
 
-export type CliJsonValue =
-  | null
-  | boolean
-  | number
-  | string
-  | readonly CliJsonValue[]
-  | { readonly [key: string]: CliJsonValue };
-
-export interface CliTarget {
-  readonly command: string;
-  readonly args: readonly string[];
-  readonly cwd?: string;
-  readonly env?: Readonly<Record<string, string>>;
-}
-
-export interface CliCapabilitySummary {
-  readonly id: string;
-  readonly description: string;
-  readonly title?: string;
-  readonly annotations?: Readonly<Record<string, CliJsonValue>>;
-}
-
-export interface CliCapabilityDescription extends CliCapabilitySummary {
-  readonly inputSchema: Readonly<Record<string, CliJsonValue>>;
-  readonly outputSchema: Readonly<Record<string, CliJsonValue>>;
-  readonly timeoutMs?: number;
-}
-
-export interface CliConnectionSummary {
-  readonly command: string;
-  readonly capabilityCount: number;
-}
-
-export type CliConnectionState =
-  | {
-      readonly state: "idle";
-      readonly validation?: {
-        readonly status: "error";
-        readonly error: { readonly code: string; readonly message: string };
-      };
-      readonly activity?: readonly CliActivityRecord[];
-    }
-  | { readonly state: "busy" | "connecting" | "closing" }
-  | {
-      readonly state: "connected";
-      readonly connection?: CliConnectionSummary;
-    };
-
-export type CliSession = CliConnectionState & {
-  readonly csrfToken: string;
-};
-
-export interface CliActivityRecord {
-  readonly sequence: number | string;
-  readonly operation: "list" | "describe" | "run" | "disconnect";
-  readonly startedAt: string;
-  readonly durationMs: number;
-  readonly outcome: string;
-  readonly errorCode?: string;
-  readonly capabilityId?: string;
-  readonly exitCode?: number | null;
-}
-
-export interface CliApi {
-  session(): Promise<CliSession>;
-  connect(target: CliTarget): Promise<CliConnectionState>;
-  disconnect(): Promise<CliConnectionState>;
-  refresh(): Promise<CliConnectionState>;
-  catalog(): Promise<readonly CliCapabilitySummary[]>;
-  describe(id: string): Promise<CliCapabilityDescription>;
-  run(id: string, input: CliJsonValue): Promise<CliJsonValue>;
-  activity(): Promise<readonly CliActivityRecord[]>;
-}
-
-export interface CliTargetDraft {
-  readonly command: string;
-  readonly args: readonly string[];
-  readonly cwd?: string;
-  readonly environment: readonly Readonly<{
-    name: string;
-    value: string;
-  }>[];
-}
-
-export interface SecretControl {
-  value: string;
-  placeholder?: string;
-}
+export type {
+  CliActivityRecord,
+  CliApi,
+  CliCapabilityDescription,
+  CliCapabilitySummary,
+  CliConnectionState,
+  CliConnectionSummary,
+  CliJsonValue,
+  CliSession,
+  CliTarget,
+  CliTargetDraft,
+  SecretControl,
+} from "./cli-contract.js";
+export {
+  buildCliTarget,
+  completeConnectionAttempt,
+  nextRovingIndex,
+  parseRunInput,
+  refreshFailureIsDisconnect,
+  retainedActivityOf,
+  seedCliInput,
+} from "./cli-contract.js";
+export { CliApiError, createRouteCliApi } from "./cli-api.js";
 
 export const cliPrimaryTabs = ["Commands", "Activity", "Connection"] as const;
 
-const activityPollDelayMs = 5_000;
-const environmentNamePattern = /^[A-Za-z_][A-Za-z0-9_]*$/;
-
 type CliTab = "commands" | "activity" | "connection";
-type Fetcher = typeof fetch;
-type RovingOrientation = "horizontal" | "vertical" | "both";
-type TargetDraftField = "command" | "environment-name" | "environment-value";
 
-class TargetDraftValidationError extends Error {
-  readonly field: TargetDraftField;
-  readonly index: number | undefined;
-
-  constructor(field: TargetDraftField, message: string, index?: number) {
-    super(message);
-    this.name = "TargetDraftValidationError";
-    this.field = field;
-    this.index = index;
-  }
-}
-
-export function nextRovingIndex(
-  current: number,
-  itemCount: number,
-  key: string,
-  orientation: RovingOrientation,
-): number | undefined {
-  if (itemCount <= 0 || current < 0 || current >= itemCount) return undefined;
-  let next = current;
-  if (
-    (orientation === "horizontal" || orientation === "both") &&
-    key === "ArrowRight"
-  ) {
-    next += 1;
-  } else if (
-    (orientation === "horizontal" || orientation === "both") &&
-    key === "ArrowLeft"
-  ) {
-    next -= 1;
-  } else if (
-    (orientation === "vertical" || orientation === "both") &&
-    key === "ArrowDown"
-  ) {
-    next += 1;
-  } else if (
-    (orientation === "vertical" || orientation === "both") &&
-    key === "ArrowUp"
-  ) {
-    next -= 1;
-  } else if (key === "Home") {
-    return 0;
-  } else if (key === "End") {
-    return itemCount - 1;
-  } else {
-    return undefined;
-  }
-  return (next + itemCount) % itemCount;
-}
-
-function nonEmpty(value: string, message: string): string {
-  const normalized = value.trim();
-  if (normalized === "") {
-    throw new TargetDraftValidationError("command", message);
-  }
-  return normalized;
-}
-
-export function buildCliTarget(draft: CliTargetDraft): CliTarget {
-  const command = nonEmpty(draft.command, "Command is required.");
-  const cwd = draft.cwd?.trim();
-  const env = Object.create(null) as Record<string, string>;
-
-  for (const [index, entry] of draft.environment.entries()) {
-    const name = entry.name.trim();
-    if (!environmentNamePattern.test(name)) {
-      throw new TargetDraftValidationError(
-        "environment-name",
-        "Environment names must use letters, numbers, and underscores.",
-        index,
-      );
-    }
-    if (Object.hasOwn(env, name)) {
-      throw new TargetDraftValidationError(
-        "environment-name",
-        "Environment names must be unique.",
-        index,
-      );
-    }
-    if (entry.value === "") {
-      throw new TargetDraftValidationError(
-        "environment-value",
-        "Environment values cannot be empty.",
-        index,
-      );
-    }
-    Object.defineProperty(env, name, {
-      configurable: true,
-      enumerable: true,
-      value: entry.value,
-      writable: true,
-    });
-  }
-
-  return {
-    command,
-    args: [...draft.args],
-    ...(cwd === undefined || cwd === "" ? {} : { cwd }),
-    ...(Object.keys(env).length === 0 ? {} : { env }),
-  };
-}
-
-function clearSecrets(controls: readonly SecretControl[]): void {
-  for (const control of controls) {
-    control.value = "";
-    control.placeholder = "Cleared after response";
-  }
-}
-
-export async function completeConnectionAttempt<Value>(
-  attempt: Promise<Value>,
-  secretControls: readonly SecretControl[],
-): Promise<Value> {
-  try {
-    return await attempt;
-  } finally {
-    clearSecrets(secretControls);
-  }
-}
-
-export function seedCliInput(
-  schema: Readonly<Record<string, CliJsonValue>>,
-): string {
-  const example = exampleFromSchema(schema);
-  if (typeof example !== "object" || example === null || Array.isArray(example))
-    return "{}";
-  return pretty(example);
-}
-
-export function parseRunInput(
-  source: string,
-): Readonly<Record<string, CliJsonValue>> {
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(source);
-  } catch {
-    throw new Error("Run input must be valid JSON.");
-  }
-  if (typeof parsed !== "object" || parsed === null || Array.isArray(parsed)) {
-    throw new Error("Run input must be a JSON object.");
-  }
-  return parsed as Readonly<Record<string, CliJsonValue>>;
-}
-
-export function retainedActivityOf(
-  state: CliConnectionState,
-): readonly CliActivityRecord[] {
-  if (state.state !== "idle") return [];
-  return Array.isArray(state.activity) ? state.activity : [];
-}
-
-export class CliApiError extends Error {
-  readonly code: string;
-
-  constructor(code: string, message: string) {
-    super(message);
-    this.name = "CliApiError";
-    this.code = code;
-  }
-}
-
-/** Refresh is another `list`. Any list failure except busy disconnects. */
-export function refreshFailureIsDisconnect(code: string | undefined): boolean {
-  return code !== "TARGET_BUSY";
-}
-
-async function responseJson<Value>(response: Response): Promise<Value> {
-  let value: unknown;
-  try {
-    value = await response.json();
-  } catch {
-    throw new CliApiError(
-      "PROTOCOL_ERROR",
-      "The local interface returned an invalid response.",
-    );
-  }
-  if (!response.ok) {
-    const error = value as {
-      readonly code?: unknown;
-      readonly message?: unknown;
-    };
-    throw new CliApiError(
-      typeof error.code === "string" ? error.code : "CONNECTION_FAILED",
-      typeof error.message === "string"
-        ? error.message
-        : "The local request could not be completed.",
-    );
-  }
-  return value as Value;
-}
-
-export function createRouteCliApi(fetcher: Fetcher = fetch): CliApi {
-  let csrfToken = "";
-
-  const get = async <Value>(path: string): Promise<Value> => {
-    const response = await fetcher(path, { credentials: "same-origin" });
-    return responseJson<Value>(response);
-  };
-
-  const mutate = async <Value>(
-    path: string,
-    method: "POST" | "DELETE",
-    body?: unknown,
-  ): Promise<Value> => {
-    const response = await fetcher(path, {
-      method,
-      credentials: "same-origin",
-      headers: {
-        ...(body === undefined ? {} : { "content-type": "application/json" }),
-        "X-Invokta-CSRF": csrfToken,
-      },
-      ...(body === undefined ? {} : { body: JSON.stringify(body) }),
-    });
-    const value = await responseJson<Value>(response);
-    const replacement = response.headers.get("X-Invokta-CSRF");
-    if (replacement !== null && replacement !== "") csrfToken = replacement;
-    return value;
-  };
-
-  return {
-    async session() {
-      const response = await get<CliSession>("/api/session");
-      csrfToken = response.csrfToken;
-      return response;
-    },
-    connect: (target) =>
-      mutate<CliConnectionState>("/api/connection", "POST", target),
-    disconnect: () => mutate<CliConnectionState>("/api/connection", "DELETE"),
-    refresh: () => mutate<CliConnectionState>("/api/refresh", "POST"),
-    async catalog() {
-      const response = await get<{
-        readonly capabilities: readonly CliCapabilitySummary[];
-      }>("/api/catalog");
-      return response.capabilities;
-    },
-    describe: (id) =>
-      mutate<CliCapabilityDescription>("/api/describe", "POST", { id }),
-    async run(id, input) {
-      const response = await mutate<{ readonly result: CliJsonValue }>(
-        "/api/run",
-        "POST",
-        { id, input },
-      );
-      return response.result;
-    },
-    async activity() {
-      const response = await get<{
-        readonly records: readonly CliActivityRecord[];
-      }>("/api/activity");
-      return response.records;
-    },
-  };
-}
-
-function createBrandMark(): SVGSVGElement {
-  const namespace = "http://www.w3.org/2000/svg";
-  const mark = document.createElementNS(namespace, "svg");
-  mark.setAttribute("viewBox", "0 0 51 43");
-  mark.setAttribute("width", "24");
-  mark.setAttribute("height", "20");
-  mark.setAttribute("fill", "none");
-  mark.setAttribute("class", "att-brand-mark");
-  mark.setAttribute("aria-hidden", "true");
-  mark.setAttribute("focusable", "false");
-  const strokes = document.createElementNS(namespace, "g");
-  strokes.setAttribute("transform", "translate(-6.5,-10.5)");
-  strokes.setAttribute("stroke-width", "9");
-  strokes.setAttribute("stroke-linecap", "round");
-  strokes.setAttribute("stroke-linejoin", "round");
-  const prompt = document.createElementNS(namespace, "path");
-  prompt.setAttribute("d", "M11 15 L29 32 L11 49");
-  prompt.setAttribute("stroke", "var(--att-accent-text)");
-  const cursor = document.createElementNS(namespace, "path");
-  cursor.setAttribute("d", "M36 49 H53");
-  cursor.setAttribute("stroke", "var(--att-fg)");
-  strokes.append(prompt, cursor);
-  mark.append(strokes);
-  return mark;
-}
-
-let controlSequence = 0;
-
-function controlId(name: string): string {
-  controlSequence += 1;
-  return `cli-${name}-${String(controlSequence)}`;
-}
-
-function labelFor(
-  text: string,
-  id: string,
-  className = "att-label",
-): HTMLLabelElement {
-  return el("label", { for: id, class: className }, [text]);
-}
-
-function textInput(options: {
-  readonly label: string;
-  readonly name: string;
-  readonly value?: string;
-  readonly placeholder?: string;
-  readonly type?: "text" | "search" | "password";
-  readonly autocomplete?: string;
-}): { readonly field: HTMLDivElement; readonly input: HTMLInputElement } {
-  const id = controlId(options.name);
-  const input = el("input", {
-    id,
-    class: "att-input",
-    type: options.type ?? "text",
-    ...(options.placeholder === undefined
-      ? {}
-      : { placeholder: options.placeholder }),
-    ...(options.autocomplete === undefined
-      ? {}
-      : { autocomplete: options.autocomplete }),
-  });
-  input.value = options.value ?? "";
-  return {
-    field: el("div", {}, [labelFor(options.label, id), input]),
-    input,
-  };
-}
-
-function errorMessage(error: unknown): string {
-  return error instanceof Error
-    ? error.message
-    : "The local request could not be completed.";
-}
-
-const annotationLabels: Readonly<Record<string, string>> = {
-  readOnly: "Read-only",
-  destructive: "Destructive",
-  idempotent: "Idempotent",
-  openWorld: "Open world",
-};
-
-function annotationTags(
-  annotations: Readonly<Record<string, CliJsonValue>> | undefined,
-): readonly HTMLElement[] {
-  return Object.entries(annotations ?? {})
-    .filter(([, value]) => value === true)
-    .map(([name]) =>
-      el(
-        "span",
-        {
-          class: `att-tag${name === "destructive" ? " danger" : ""}`,
-        },
-        [annotationLabels[name] ?? name],
-      ),
-    );
-}
-
-function clockReading(instant: string): string {
-  const separator = instant.indexOf("T");
-  if (separator === -1) return instant;
-  return instant.slice(separator + 1).replace("Z", "");
-}
-
-function activityTable(records: readonly CliActivityRecord[]): HTMLElement {
-  const body = el(
-    "tbody",
-    {},
-    records.map((record) =>
-      el("tr", {}, [
-        el("td", { class: "att-mono" }, [String(record.sequence)]),
-        el("td", {}, [record.operation]),
-        el("td", { class: "att-mono" }, [record.capabilityId ?? "—"]),
-        el("td", { class: "att-mono att-numeric" }, [
-          record.exitCode === undefined || record.exitCode === null
-            ? "—"
-            : String(record.exitCode),
-        ]),
-        el(
-          "td",
-          {
-            class: "att-mono att-numeric",
-            title: clockReading(record.startedAt),
-          },
-          [`${String(record.durationMs)} ms`],
-        ),
-        el(
-          "td",
-          {
-            class: `att-status-text ${record.outcome === "success" ? "success" : "error"}`,
-          },
-          [record.errorCode ?? record.outcome],
-        ),
-      ]),
-    ),
-  );
-  return el("table", { class: "att-activity-table" }, [
-    el("thead", {}, [
-      el("tr", {}, [
-        el("th", { scope: "col" }, ["Sequence"]),
-        el("th", { scope: "col" }, ["Verb"]),
-        el("th", { scope: "col" }, ["Id"]),
-        el("th", { scope: "col" }, ["Exit"]),
-        el("th", { scope: "col" }, ["Duration"]),
-        el("th", { scope: "col" }, ["Outcome"]),
-      ]),
-    ]),
-    body,
-  ]);
-}
-
-function statusPill(state: CliConnectionState): HTMLElement {
-  const label =
-    state.state === "connected"
-      ? "Connected"
-      : state.state === "idle"
-        ? "Not connected"
-        : state.state === "busy"
-          ? "Target busy"
-          : state.state === "connecting"
-            ? "Connecting"
-            : "Closing";
-  const tone =
-    state.state === "connected"
-      ? "success"
-      : state.state === "idle"
-        ? ""
-        : "warning";
-  return el(
-    "span",
-    {
-      class: `att-pill ${tone}`.trim(),
-      role: "status",
-      "aria-live": "polite",
-    },
-    [el("span", { class: "att-dot", "aria-hidden": "true" }, []), label],
-  );
-}
+const activityPollDelayMs = 5_000;
 
 export interface CliAppHandle {
   destroy(): void;
@@ -542,7 +64,6 @@ export function mountCliApp(
   root: HTMLElement,
   api: CliApi = createRouteCliApi(),
 ): CliAppHandle {
-  type EditablePair = SecretControl & { name: string };
   const ownerDocument = root.ownerDocument;
   ownerDocument.body.classList.add("attached-mode");
   if (
@@ -575,29 +96,48 @@ export function mountCliApp(
   let activityPoll: ReturnType<typeof setInterval> | undefined;
   let connectionError = "";
   let secretConfigured = false;
-  const draft: {
-    command: string;
-    args: string[];
-    cwd: string;
-    environment: EditablePair[];
-  } = { command: "", args: [], cwd: "", environment: [] };
+  const draft: EditableCliTargetDraft = {
+    command: "",
+    args: [],
+    cwd: "",
+    environment: [],
+  };
   const fullThemeToggle = createThemeToggle();
   const compactThemeToggle = createCompactThemeToggle();
   const argumentDrafts = new Map<string, string>();
-  let currentResult:
-    | { readonly id: string; readonly value: CliJsonValue }
-    | undefined;
+  let currentResult: CliCurrentResult | undefined;
 
   const render = (): void => {
     if (destroyed) return;
-    if (targetState.state === "idle") renderIdle();
-    else if (
+    if (targetState.state === "idle") {
+      shell(
+        createCliIdleView({
+          state: targetState,
+          draft,
+          connectionError,
+          connect: (target) => api.connect(target),
+          connected(state, carriesSecrets) {
+            targetState = state;
+            secretConfigured = carriesSecrets;
+            connectionError = "";
+            render();
+            if (
+              targetState.state === "connected" &&
+              targetState.connection !== undefined
+            ) {
+              void loadCatalog();
+            }
+          },
+        }),
+        false,
+      );
+    } else if (
       targetState.state === "connected" &&
       targetState.connection !== undefined
     ) {
       renderConnected();
     } else {
-      renderUnavailableState();
+      shell(createCliUnavailableView(targetState), false);
     }
   };
 
@@ -656,16 +196,17 @@ export function mountCliApp(
 
   const shell = (content: HTMLElement, includeTabs: boolean): void => {
     const brand = el("div", { class: "att-brand" }, [
-      createBrandMark(),
+      createCliBrandMark(),
       el("span", { class: "att-brand-name" }, ["invokta"]),
       el("span", { class: "att-product-name" }, ["DevTools"]),
     ]);
     const topbarChildren: Node[] = [brand];
     if (includeTabs) topbarChildren.push(createTabs());
-    else
+    else {
       topbarChildren.push(
         el("span", { class: "att-topbar-spacer", "aria-hidden": "true" }, []),
       );
+    }
     topbarChildren.push(
       el("div", { class: "att-theme-slot att-theme-slot-full" }, [
         fullThemeToggle,
@@ -693,7 +234,7 @@ export function mountCliApp(
           el("h1", {}, [heading]),
         ]),
         el("div", { class: "att-context-meta" }, [
-          statusPill(targetState),
+          cliStatusPill(targetState),
           ...(targetState.state === "connected" &&
           targetState.connection !== undefined
             ? [
@@ -711,708 +252,51 @@ export function mountCliApp(
     );
   };
 
-  function renderIdle(): void {
-    const intro = el("section", { class: "att-idle-intro" }, [
-      el("p", { class: "att-kicker" }, ["Connection"]),
-      el("h2", {}, ["Attach one Invokta CLI"]),
-      el("p", {}, [
-        "Name the executable, arguments, working directory, and environment the operator will type.",
-      ]),
-      el("p", { class: "att-hint" }, [
-        "Nothing is discovered, imported, invoked, or saved automatically.",
-      ]),
-    ]);
-    const formHost = el("section", {
-      class: "att-idle-form",
-      "aria-label": "Connection details",
-    });
-
-    const paintForm = (): void => {
-      clear(formHost);
-      const validationFields = new Map<string, HTMLInputElement>();
-      const validationKey = (field: TargetDraftField, index?: number): string =>
-        `${field}:${index === undefined ? "single" : String(index)}`;
-      const feedback = el("p", {
-        id: controlId("connection-feedback"),
-        class: "att-feedback",
-        role: "alert",
-        "aria-live": "assertive",
-      });
-      feedback.textContent = connectionError;
-      const registerValidationField = (
-        field: TargetDraftField,
-        input: HTMLInputElement,
-        index?: number,
-      ): void => {
-        validationFields.set(validationKey(field, index), input);
-        input.addEventListener("input", () => {
-          if (input.getAttribute("aria-invalid") !== "true") return;
-          input.removeAttribute("aria-invalid");
-          input.removeAttribute("aria-errormessage");
-          feedback.textContent = "";
-        });
-      };
-
-      const form = el("form", {}, []);
-      const command = textInput({
-        label: "Command",
-        name: "command",
-        value: draft.command,
-        placeholder: "node",
-        autocomplete: "off",
-      });
-      command.input.required = true;
-      command.input.addEventListener("input", () => {
-        draft.command = command.input.value;
-      });
-      registerValidationField("command", command.input);
-      const cwd = textInput({
-        label: "Working directory",
-        name: "cwd",
-        value: draft.cwd,
-        placeholder: "Optional directory",
-        autocomplete: "off",
-      });
-      cwd.input.addEventListener("input", () => {
-        draft.cwd = cwd.input.value;
-      });
-      const argumentsHost = el("div", {}, []);
-      const addArgumentRows = (
-        host: HTMLElement,
-        fallbackFocus?: HTMLElement,
-        focusIndex?: number,
-      ): void => {
-        clear(host);
-        for (const [index, argument] of draft.args.entries()) {
-          const id = controlId("argument");
-          const input = el("input", {
-            id,
-            class: "att-input",
-            type: "text",
-            "aria-label": `Argument ${String(index + 1)}`,
-          });
-          input.value = argument;
-          input.addEventListener("input", () => {
-            draft.args[index] = input.value;
-          });
-          const remove = el(
-            "button",
-            {
-              type: "button",
-              class: "att-button att-icon-button",
-              "aria-label": `Remove argument ${String(index + 1)}`,
-            },
-            ["Remove"],
-          );
-          remove.addEventListener("click", () => {
-            draft.args.splice(index, 1);
-            const nextIndex = Math.min(index, draft.args.length - 1);
-            addArgumentRows(
-              host,
-              fallbackFocus,
-              nextIndex >= 0 ? nextIndex : undefined,
-            );
-            if (nextIndex < 0) fallbackFocus?.focus();
-          });
-          host.append(
-            el("div", { class: "att-inline-fields single" }, [
-              el("div", {}, [
-                labelFor(`Argument ${String(index + 1)}`, id),
-                input,
-              ]),
-              remove,
-            ]),
-          );
-        }
-        if (focusIndex !== undefined) {
-          host
-            .querySelector<HTMLInputElement>(
-              `[aria-label="Argument ${String(focusIndex + 1)}"]`,
-            )
-            ?.focus();
-        }
-      };
-      const addArgument = el(
-        "button",
-        { type: "button", class: "att-button" },
-        ["Add argument"],
-      );
-      addArgument.addEventListener("click", () => {
-        draft.args.push("");
-        addArgumentRows(argumentsHost, addArgument, draft.args.length - 1);
-      });
-      addArgumentRows(argumentsHost, addArgument);
-
-      const environmentHost = el("div", {}, []);
-      const addPairRows = (
-        host: HTMLElement,
-        pairs: EditablePair[],
-        fallbackFocus?: HTMLElement,
-        focusIndex?: number,
-      ): void => {
-        clear(host);
-        for (const [index, pair] of pairs.entries()) {
-          const nameField = textInput({
-            label: "Variable name",
-            name: "environment-name",
-            value: pair.name,
-            placeholder: "API_TOKEN",
-          });
-          const valueField = textInput({
-            label: "Variable value",
-            name: "environment-value",
-            value: pair.value,
-            ...(pair.placeholder === undefined
-              ? {}
-              : { placeholder: pair.placeholder }),
-            type: "password",
-            autocomplete: "off",
-          });
-          nameField.input.addEventListener("input", () => {
-            pair.name = nameField.input.value;
-          });
-          valueField.input.addEventListener("input", () => {
-            pair.value = valueField.input.value;
-          });
-          registerValidationField("environment-name", nameField.input, index);
-          registerValidationField("environment-value", valueField.input, index);
-          const remove = el(
-            "button",
-            {
-              type: "button",
-              class: "att-button att-icon-button",
-              "aria-label": `Remove environment row ${String(index + 1)}`,
-            },
-            ["Remove"],
-          );
-          remove.addEventListener("click", () => {
-            clearSecrets([pair, valueField.input]);
-            pairs.splice(index, 1);
-            const nextIndex = Math.min(index, pairs.length - 1);
-            addPairRows(
-              host,
-              pairs,
-              fallbackFocus,
-              nextIndex >= 0 ? nextIndex : undefined,
-            );
-            if (nextIndex < 0) fallbackFocus?.focus();
-          });
-          host.append(
-            el("div", { class: "att-inline-fields" }, [
-              nameField.field,
-              valueField.field,
-              remove,
-            ]),
-          );
-        }
-        if (focusIndex !== undefined) {
-          validationFields
-            .get(validationKey("environment-name", focusIndex))
-            ?.focus();
-        }
-      };
-      const addEnvironment = el(
-        "button",
-        { type: "button", class: "att-button" },
-        ["Add environment variable"],
-      );
-      addEnvironment.addEventListener("click", () => {
-        draft.environment.push({ name: "", value: "", placeholder: "" });
-        addPairRows(
-          environmentHost,
-          draft.environment,
-          addEnvironment,
-          draft.environment.length - 1,
-        );
-      });
-      addPairRows(environmentHost, draft.environment, addEnvironment);
-
-      const connect = el(
-        "button",
-        { type: "submit", class: "att-button primary" },
-        ["Connect"],
-      );
-      form.append(
-        command.field,
-        cwd.field,
-        el("div", { class: "att-section-heading" }, [
-          el("h3", {}, ["Arguments"]),
-          el("span", { class: "att-hint" }, ["One exact value per row"]),
-        ]),
-        argumentsHost,
-        addArgument,
-        el("div", { class: "att-section-heading" }, [
-          el("h3", {}, ["Environment"]),
-          el("span", { class: "att-hint" }, ["Values stay in memory"]),
-        ]),
-        environmentHost,
-        addEnvironment,
-        el("div", { class: "att-actions" }, [connect]),
-        feedback,
-      );
-      form.addEventListener("submit", (event) => {
-        event.preventDefault();
-        feedback.textContent = "";
-        let target: CliTarget;
-        try {
-          target = buildCliTarget({
-            command: draft.command,
-            args: draft.args,
-            cwd: draft.cwd,
-            environment: draft.environment,
-          });
-        } catch (error) {
-          feedback.textContent = errorMessage(error);
-          if (error instanceof TargetDraftValidationError) {
-            validationFields
-              .get(validationKey(error.field, error.index))
-              ?.setAttribute("aria-invalid", "true");
-          }
-          return;
-        }
-        const secretControls = [...draft.environment];
-        const passwordInputs = Array.from(
-          form.querySelectorAll<HTMLInputElement>('input[type="password"]'),
-        );
-        const carriesSecrets = secretControls.length > 0;
-        connect.disabled = true;
-        connect.textContent = "Connecting…";
-        void completeConnectionAttempt(api.connect(target), [
-          ...secretControls,
-          ...passwordInputs,
-        ])
-          .then((state) => {
-            targetState = state;
-            secretConfigured = carriesSecrets;
-            connectionError = "";
-            render();
-            if (
-              targetState.state === "connected" &&
-              targetState.connection !== undefined
-            ) {
-              void loadCatalog();
-            }
-          })
-          .catch((error: unknown) => {
-            connect.disabled = false;
-            connect.textContent = "Connect";
-            feedback.textContent = errorMessage(error);
-          });
-      });
-
-      const orientation = el("aside", { class: "att-idle-orient" }, [
-        el("h3", {}, ["This is the CLI workbench."]),
-        el("p", {}, [
-          "It inspects an installed Invokta CLI without loading an engine.",
-        ]),
-        el("dl", { class: "att-idle-orient-paths" }, [
-          el("dt", {}, ["MCP workbench"]),
-          el("dd", {}, [
-            el("div", { class: "att-mono" }, ["invokta-devtools"]),
-            el("div", {}, ["or invokta-devtools open"]),
-          ]),
-          el("dt", {}, ["Project workspace"]),
-          el("dd", {}, [
-            el("div", { class: "att-mono" }, [
-              "invokta-devtools serve dist/engine.js",
-            ]),
-            el("div", {}, ["or yarn devtools inside an engine repo"]),
-          ]),
-        ]),
-      ]);
-      formHost.append(form, orientation);
-    };
-
-    paintForm();
-    const card = el("div", { class: "att-card att-idle" }, [intro, formHost]);
-    const retained = retainedActivityOf(targetState);
-    if (retained.length === 0) {
-      shell(card, false);
-      return;
-    }
-    shell(
-      el("div", {}, [
-        card,
-        el(
-          "section",
-          {
-            class: "att-card att-view att-retained",
-            "aria-label": "Activity retained from the disconnected target",
-          },
-          [
-            el("div", { class: "att-section-heading" }, [
-              el("div", {}, [
-                el("h2", {}, ["Last session activity"]),
-                el("span", { class: "att-hint" }, [
-                  "CLI verbs retained from the disconnected target",
-                ]),
-              ]),
-            ]),
-            activityTable(retained),
-          ],
-        ),
-      ]),
-      false,
-    );
-  }
-
-  function renderUnavailableState(): void {
-    const label =
-      targetState.state === "connecting"
-        ? "A connection is being established."
-        : targetState.state === "closing"
-          ? "The current connection is closing."
-          : "A target is connected in another local browser session.";
-    shell(
-      el("section", { class: "att-card att-view" }, [
-        el("p", { class: "att-kicker" }, ["Connection"]),
-        el("h2", {}, ["Target slot unavailable"]),
-        el("p", { class: "att-hint" }, [label]),
-      ]),
-      false,
-    );
-  }
-
   function renderConnected(): void {
     let panel: HTMLElement;
     if (activeTab === "activity") panel = renderActivityPanel();
     else if (activeTab === "connection") panel = renderConnectionPanel();
-    else panel = renderCommandsPanel();
+    else {
+      panel = createCliCommandsPanel({
+        root,
+        api,
+        commands,
+        commandsLoaded,
+        commandsLoading,
+        commandsError,
+        selectedId,
+        described,
+        describeError,
+        commandQuery,
+        argumentDrafts,
+        currentResult,
+        retryCatalog: () => void loadCatalog(),
+        setSelectedId: (id) => {
+          selectedId = id;
+        },
+        selectCommand(id) {
+          selectedId = id;
+          described = undefined;
+          describeError = "";
+          render();
+          void loadDescribe(id);
+        },
+        setCommandQuery: (query) => {
+          commandQuery = query;
+        },
+        setCurrentResult: (value) => {
+          currentResult = value;
+        },
+        markActivityStale: () => {
+          activityLoaded = false;
+        },
+      });
+    }
     panel.id = `cli-panel-${activeTab}`;
     panel.setAttribute("role", "tabpanel");
     panel.setAttribute("aria-labelledby", `cli-tab-${activeTab}`);
     panel.setAttribute("tabindex", "0");
     shell(el("div", { class: "att-card att-workbench" }, [panel]), true);
-  }
-
-  function filterCommands(
-    items: readonly CliCapabilitySummary[],
-    query: string,
-  ): readonly CliCapabilitySummary[] {
-    const needle = query.trim().toLowerCase();
-    if (needle === "") return items;
-    return items.filter((item) =>
-      [item.id, item.title ?? "", item.description].some((value) =>
-        value.toLowerCase().includes(needle),
-      ),
-    );
-  }
-
-  function renderCommandsPanel(): HTMLElement {
-    if (commandsLoading && !commandsLoaded) {
-      return el("section", { class: "att-loading" }, ["Loading commands…"]);
-    }
-    if (commandsError !== "" && !commandsLoaded) {
-      const retry = el("button", { type: "button", class: "att-button" }, [
-        "Refresh catalog",
-      ]);
-      retry.addEventListener("click", () => void loadCatalog());
-      return el("section", { class: "att-view" }, [
-        el("h2", {}, ["Commands unavailable"]),
-        el("p", { class: "att-feedback", role: "alert" }, [commandsError]),
-        retry,
-      ]);
-    }
-
-    if (
-      selectedId === undefined ||
-      !commands.some(({ id }) => id === selectedId)
-    ) {
-      selectedId = commands[0]?.id;
-    }
-    const sidebar = el("aside", { class: "att-tool-sidebar" }, [
-      el("h2", {}, ["Commands"]),
-      el("p", { class: "att-hint" }, [
-        `${String(commands.length)} advertised commands`,
-      ]),
-    ]);
-    const search = textInput({
-      label: "Search commands",
-      name: "command-search",
-      value: commandQuery,
-      placeholder: "Id, title, or description",
-      type: "search",
-      autocomplete: "off",
-    });
-    const list = el(
-      "div",
-      {
-        class: "att-tool-list",
-        role: "listbox",
-        "aria-label": "Connected CLI commands",
-      },
-      [],
-    );
-    let visible: readonly CliCapabilitySummary[] = [];
-    let choiceButtons: HTMLButtonElement[] = [];
-    const focusChoice = (id: string): void => {
-      Array.from(root.querySelectorAll<HTMLButtonElement>(".att-tool-choice"))
-        .find((candidate) => candidate.dataset.commandId === id)
-        ?.focus();
-    };
-    const selectCommand = (id: string, focus: boolean): void => {
-      selectedId = id;
-      described = undefined;
-      describeError = "";
-      render();
-      void loadDescribe(id);
-      if (focus) focusChoice(id);
-    };
-    const paintCommands = (): void => {
-      clear(list);
-      const matches = filterCommands(commands, search.input.value);
-      visible = matches;
-      choiceButtons = [];
-      if (matches.length === 0) {
-        list.append(el("p", { class: "att-empty" }, ["No matching commands."]));
-        return;
-      }
-      const rovingId = matches.some(({ id }) => id === selectedId)
-        ? selectedId
-        : matches[0]?.id;
-      for (const item of matches) {
-        const choice = el(
-          "button",
-          {
-            type: "button",
-            class: "att-tool-choice",
-            role: "option",
-            "aria-selected": String(item.id === selectedId),
-            tabindex: item.id === rovingId ? "0" : "-1",
-            "data-command-id": item.id,
-          },
-          [
-            el("span", { class: "att-tool-title" }, [item.title ?? item.id]),
-            el("span", { class: "att-tool-name" }, [item.id]),
-          ],
-        );
-        choice.addEventListener("click", () => {
-          selectCommand(item.id, true);
-        });
-        choiceButtons.push(choice);
-        list.append(choice);
-      }
-    };
-    search.input.addEventListener("input", () => {
-      commandQuery = search.input.value;
-      paintCommands();
-    });
-    list.addEventListener("keydown", (event) => {
-      const current = choiceButtons.indexOf(event.target as HTMLButtonElement);
-      const next = nextRovingIndex(
-        current,
-        choiceButtons.length,
-        event.key,
-        "vertical",
-      );
-      if (next === undefined) return;
-      event.preventDefault();
-      const item = visible[next];
-      if (item !== undefined) selectCommand(item.id, true);
-    });
-    sidebar.append(search.field, list);
-    paintCommands();
-
-    const selected = commands.find(({ id }) => id === selectedId);
-    if (selected === undefined) {
-      return el("div", { class: "att-tools-layout" }, [
-        sidebar,
-        el("section", { class: "att-tool-detail" }, [
-          el("div", { class: "att-empty" }, [
-            "The connected CLI did not advertise any commands.",
-          ]),
-        ]),
-      ]);
-    }
-
-    const contractCopy = el("p", { class: "att-hint" }, [
-      "This CLI uses its own composition root. DevTools does not supply a principal. Each verb starts a new process and the process exits.",
-    ]);
-
-    if (described === undefined || described.id !== selected.id) {
-      return el("div", { class: "att-tools-layout" }, [
-        sidebar,
-        el("section", { class: "att-tool-detail" }, [
-          el("header", { class: "att-tool-header" }, [
-            el("h2", {}, [selected.title ?? selected.id]),
-            el("span", { class: "att-tool-name" }, [selected.id]),
-          ]),
-          contractCopy,
-          describeError === ""
-            ? el("p", { class: "att-hint" }, ["Reading the describe contract…"])
-            : el("p", { class: "att-feedback", role: "alert" }, [
-                describeError,
-              ]),
-        ]),
-      ]);
-    }
-
-    const selectedDescription = described;
-    const schemaText = pretty(selectedDescription.inputSchema);
-    const schema = el("pre", { class: "att-pre" }, [schemaText]);
-    const seed = seedCliInput(selectedDescription.inputSchema);
-    const source = argumentDrafts.get(selectedDescription.id) ?? seed;
-    const argumentsId = controlId("arguments");
-    const argumentsEditor = el("textarea", {
-      id: argumentsId,
-      class: "att-textarea",
-      spellcheck: "false",
-      "aria-label": `JSON input for ${selectedDescription.id}`,
-    });
-    argumentsEditor.value = source;
-    argumentsEditor.addEventListener("input", () => {
-      argumentDrafts.set(selectedDescription.id, argumentsEditor.value);
-    });
-    const feedback = el("p", {
-      class: "att-feedback",
-      role: "alert",
-      "aria-live": "assertive",
-    });
-    const emptyResult = "Run this command to inspect its current result.";
-    const hasResult = currentResult?.id === selectedDescription.id;
-    const result = el(
-      "pre",
-      {
-        class: "att-pre att-result",
-        role: "status",
-        "aria-live": "polite",
-        "aria-label": "Current result",
-      },
-      [hasResult ? pretty(currentResult?.value) : emptyResult],
-    );
-    const resultState = el("span", {}, [hasResult ? "Returned" : "Not run"]);
-    const run = el("button", { type: "button", class: "att-button primary" }, [
-      "Run",
-    ]);
-    const format = el("button", { type: "button", class: "att-button" }, [
-      "Format JSON",
-    ]);
-    const reset = el("button", { type: "button", class: "att-button" }, [
-      "Reset to schema",
-    ]);
-    const writeDraft = (value: string): void => {
-      argumentsEditor.value = value;
-      argumentDrafts.set(selectedDescription.id, value);
-    };
-    format.addEventListener("click", () => {
-      feedback.textContent = "";
-      try {
-        writeDraft(pretty(parseRunInput(argumentsEditor.value)));
-      } catch (error) {
-        feedback.textContent = errorMessage(error);
-        argumentsEditor.focus();
-      }
-    });
-    reset.addEventListener("click", () => {
-      feedback.textContent = "";
-      writeDraft(seed);
-      argumentsEditor.focus();
-    });
-    const runCall = (): void => {
-      if (run.disabled) return;
-      feedback.textContent = "";
-      let input: Readonly<Record<string, CliJsonValue>>;
-      try {
-        input = parseRunInput(argumentsEditor.value);
-      } catch (error) {
-        feedback.textContent = errorMessage(error);
-        return;
-      }
-      run.disabled = true;
-      run.textContent = "Running…";
-      resultState.textContent = "Waiting";
-      result.textContent = "Waiting for the CLI…";
-      const startedAt = performance.now();
-      void api
-        .run(selectedDescription.id, input)
-        .then((value) => {
-          currentResult = { id: selectedDescription.id, value };
-          result.textContent = pretty(value);
-          resultState.textContent = `Returned · ${formatDuration(performance.now() - startedAt)}`;
-        })
-        .catch((error: unknown) => {
-          feedback.textContent = errorMessage(error);
-          result.textContent = "No result was returned.";
-          resultState.textContent = `Failed · ${formatDuration(performance.now() - startedAt)}`;
-        })
-        .finally(() => {
-          run.disabled = false;
-          run.textContent = "Run";
-          activityLoaded = false;
-        });
-    };
-    run.addEventListener("click", runCall);
-    argumentsEditor.addEventListener("keydown", (event) => {
-      if (event.key !== "Enter" || (!event.ctrlKey && !event.metaKey)) return;
-      event.preventDefault();
-      runCall();
-    });
-
-    const tags = annotationTags(selectedDescription.annotations);
-    const detail = el(
-      "section",
-      { id: "cli-command-detail", class: "att-tool-detail" },
-      [
-        el("header", { class: "att-tool-header" }, [
-          el("h2", {}, [selectedDescription.title ?? selectedDescription.id]),
-          el("span", { class: "att-tool-name" }, [selectedDescription.id]),
-          ...(tags.length === 0
-            ? []
-            : [el("div", { class: "att-tool-tags" }, tags)]),
-          el("p", {}, [selectedDescription.description]),
-        ]),
-        contractCopy,
-        el("div", { class: "att-tool-grid" }, [
-          el("section", { class: "att-pane", "aria-label": "Input schema" }, [
-            el("div", { class: "att-pane-bar" }, [
-              el("span", {}, ["Input schema"]),
-              el("div", { class: "att-pane-tools" }, [
-                el("span", {}, ["JSON Schema"]),
-                createCopyButton(
-                  "input schema",
-                  () => schemaText,
-                  "att-copy-button",
-                ),
-              ]),
-            ]),
-            schema,
-          ]),
-          el("section", { class: "att-pane", "aria-label": "Manual run" }, [
-            el("div", { class: "att-pane-bar" }, [
-              el("span", {}, ["Run"]),
-              el("span", {}, ["JSON"]),
-            ]),
-            el("div", { class: "att-pane-body" }, [
-              labelFor("Input", argumentsId, "att-label"),
-              argumentsEditor,
-              el("div", { class: "att-actions" }, [
-                run,
-                format,
-                reset,
-                el("span", { class: "att-shortcut" }, ["Ctrl/⌘ + Enter"]),
-              ]),
-              feedback,
-            ]),
-            el("div", { class: "att-pane-bar" }, [
-              el("span", {}, ["Current result"]),
-              el("div", { class: "att-pane-tools" }, [
-                resultState,
-                createCopyButton(
-                  "current result",
-                  () => {
-                    const text = result.textContent ?? "";
-                    return text === emptyResult ? "" : text;
-                  },
-                  "att-copy-button",
-                ),
-              ]),
-            ]),
-            result,
-          ]),
-        ]),
-      ],
-    );
-    return el("div", { class: "att-tools-layout" }, [sidebar, detail]);
   }
 
   function renderActivityPanel(): HTMLElement {
@@ -1445,7 +329,7 @@ export function mountCliApp(
         ]),
       );
     } else {
-      content.push(activityTable(activity));
+      content.push(cliActivityTable(activity));
     }
     return el("section", { class: "att-view" }, content);
   }
@@ -1500,7 +384,7 @@ export function mountCliApp(
         .catch((error: unknown) => {
           disconnect.disabled = false;
           disconnect.textContent = "Disconnect";
-          feedback.textContent = errorMessage(error);
+          feedback.textContent = cliErrorMessage(error);
         });
     });
     refresh.addEventListener("click", () => {
@@ -1518,7 +402,7 @@ export function mountCliApp(
         })
         .catch((error: unknown) => {
           const code = error instanceof CliApiError ? error.code : undefined;
-          const message = errorMessage(error);
+          const message = cliErrorMessage(error);
           if (!refreshFailureIsDisconnect(code)) {
             feedback.textContent = message;
             return;
@@ -1581,9 +465,7 @@ export function mountCliApp(
     if (name === "activity") startActivityPolling();
     else stopActivityPolling();
     render();
-    if (name === "activity") {
-      await loadActivity();
-    }
+    if (name === "activity") await loadActivity();
   }
 
   const stopActivityPolling = (): void => {
@@ -1613,7 +495,7 @@ export function mountCliApp(
       if (selectedId === undefined) selectedId = commands[0]?.id;
       if (selectedId !== undefined) void loadDescribe(selectedId);
     } catch (error) {
-      commandsError = errorMessage(error);
+      commandsError = cliErrorMessage(error);
     } finally {
       commandsLoading = false;
       render();
@@ -1626,7 +508,7 @@ export function mountCliApp(
       described = await api.describe(id);
     } catch (error) {
       described = undefined;
-      describeError = errorMessage(error);
+      describeError = cliErrorMessage(error);
     } finally {
       render();
     }
@@ -1641,7 +523,7 @@ export function mountCliApp(
       activity = await api.activity();
       activityLoaded = true;
     } catch (error) {
-      activityError = errorMessage(error);
+      activityError = cliErrorMessage(error);
     } finally {
       activityLoading = false;
       render();
@@ -1766,7 +648,7 @@ export function mountCliApp(
       targetState = { state: "idle" };
       render();
       const alert = root.querySelector<HTMLElement>(".att-feedback");
-      if (alert !== null) alert.textContent = errorMessage(error);
+      if (alert !== null) alert.textContent = cliErrorMessage(error);
     });
 
   return {
